+++ CHANGELOG.md
@@ -6,6 +6,8 @@ All notable changes to this project will be documented in this file.
 
 [Full diff](https://github.com/sider/runners/compare/0.40.6...HEAD)
 
+- Make internal changes for the `Runners::Config` class [#1950](https://github.com/sider/runners/pull/1950) [#1951](https://github.com/sider/runners/pull/1951)
+
 ## 0.40.6
 
 [Full diff](https://github.com/sider/runners/compare/0.40.5...0.40.6)